refactor: consolidate to officers pages
