+++ pages/officers/[slug].tsx
@@ -1,4 +1,4 @@
-// pages/team/[slug].js
+// pages/officers/[slug].tsx
 import { useRouter } from 'next/router';
 import Header from '@/components/Header';
 import ContactCard from '@/components/ContactCard';
